axf_VUE/requirements.txt
alembic==1.0.10
amqp==2.5.0
aniso8601==6.0.0
billiard==3.5.0.5
celery==4.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==1.11.7
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-redis==4.10.0
django-tinymce==2.8.0
djangorestframework==3.9.4
drf-extensions==0.5.0
Flask==1.0.2
Flask-Caching==1.7.1
Flask-Migrate==2.5.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.1.0
Mako==1.0.10
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.0.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
redis==2.10.6
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
urllib3==1.25.2
uWSGI==2.0.18
vine==1.3.0
Werkzeug==0.15.4
